# andes-api-toolkit/examples/poc-server/src/main/java/pe/andes/poc/server/customer/CustomerRequest.java
package pe.andes.poc.server.customer;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CustomerRequest(
        @NotBlank @Size(min = 3, max = 120) String fullName,
        @NotBlank @Email String email) {
}